requirements.txt
matplotlib>=3.0.0
Pillow>=10.0.0
numpy>=1.20.0